# backend/src/openapi.yaml
openapi: 3.0.3
info:
  title: Node JS Blog REST API
  version: 1.0.0
  description: |
    A production‐ready blog API built with Express, TypeScript, MongoDB, and JWT authentication.
    All endpoints under `/api/auth`, `/api/categories`, and `/api/posts` are documented here.
servers:
  - url: http://localhost:4000
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserResponse:
      type: object
      required:
        - id
        - username
        - email
      properties:
        id:
          type: string
          example: 60d5ec49f88a4c3e88f3ad91
        username:
          type: string
          example: johndoe
        email:
          type: string
          format: email
          example: john@example.com
        createdAt:
          type: string
          format: date-time
          example: '2025-06-02T16:15:42.405Z'
        updatedAt:
          type: string
          format: date-time
          example: '2025-06-02T16:20:12.123Z'

    #---------------------------
    # 2) Category Schemas
    #---------------------------
    Category:
      type: object
      required:
        - id
        - name
        - slug
      properties:
        id:
          type: string
          example: 60d6f1e2c9a8653c947ae16a
        name:
          type: string
          example: Technology
        slug:
          type: string
          example: technology
        createdAt:
          type: string
          format: date-time
          example: '2025-06-02T17:00:00.000Z'
        updatedAt:
          type: string
          format: date-time
          example: '2025-06-02T17:05:00.000Z'
...
    Post:
      type: object
      required:
        - id
        - title
        - slug
        - content
        - author
        - categories
      properties:
        id:
          type: string
          example: 60d70bffc9a8653c947af2b4
        title:
          type: string
          example: My First Blog Post
        slug:
          type: string
          example: my-first-blog-post
        content:
          type: string
          example: Lorem ipsum dolor sit amet...
        author:
          $ref: '#/components/schemas/UserResponse'
        categories:
          type: array
          items:
            $ref: '#/components/schemas/Category'
        imageUrl:
          type: string
          format: uri
          example: /uploads/posts/abcdef_16243512345.jpg
        createdAt:
          type: string
          format: date-time
          example: '2025-06-02T19:00:00.000Z'
        updatedAt:
          type: string
          format: date-time
          example: '2025-06-02T19:05:00.000Z'
...
    PostResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          example: true
        data:
          oneOf:
            - $ref: '#/components/schemas/Post'
            - type: array
              items:
                $ref: '#/components/schemas/Post'
    PaginatedPosts:
      type: object
      required:
        - success
        - page
        - limit
        - total
        - data
      properties:
        success:
          type: boolean
          example: true
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 10
        total:
          type: integer
          example: 42
        data:
          type: array
          items:
            $ref: '#/components/schemas/Post'

    #---------------------------
    # 4) Comment Schemas
    #---------------------------
...
    ErrorResponse:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Invalid credentials
      additionalProperties: false
paths:
...
  /api/posts:
    get:
      tags:
        - Posts
      summary: Get paginated list of posts
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: Page number
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
          description: Number of posts per page
      responses:
        '200':
          description: Paginated posts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedPosts'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Posts
      summary: Create a new post (with optional image upload)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - title
                - slug
                - content
                - authorId
                - categories
              properties:
                title:
                  type: string
                  example: My First Blog Post
                slug:
                  type: string
                  example: my-first-blog-post
                content:
                  type: string
                  example: Lorem ipsum dolor sit amet...
                authorId:
                  type: string
                  example: 60d5ec49f88a4c3e88f3ad91
                  description: MongoDB ObjectId of the author
                categories:
                  type: array
                  items:
                    type: string
                    example: 60d6f1e2c9a8653c947ae16a
                    description: MongoDB ObjectId strings of categories
                image:
                  type: string
                  format: binary
                  description: Optional JPEG/PNG file for the post’s cover image
      responses:
        '201':
          description: Post created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '400':
          description: Validation error or invalid file type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Slug conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/posts/{id}:
    get:
      tags:
        - Posts
      summary: Get a single post by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Post ID (Mongo ObjectId)
      responses:
        '200':
          description: Post found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Posts
      summary: Update a post by ID (fields + optional new image)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Post ID (Mongo ObjectId)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: My Updated Blog Post
                slug:
                  type: string
                  example: my-updated-blog-post
                content:
                  type: string
                  example: Updated content here...
                categories:
                  type: array
                  items:
                    type: string
                    example: 60d6f1e2c9a8653c947ae16a
                    description: MongoDB ObjectId strings of categories
                image:
                  type: string
                  format: binary
                  description: Optional new JPEG/PNG file to replace existing image
      responses:
        '200':
          description: Post updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '400':
          description: Validation error or invalid file type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Slug conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Posts
      summary: Delete a post by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Post ID (Mongo ObjectId)
      responses:
        '204':
          description: Post deleted (no content)
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  #---------------------------------------------
  # 4) Comment Routes (protected by Bearer JWT)
  #---------------------------------------------
...